migrate all tests
