+++ package.json
@@ -35,10 +35,10 @@
   "peerDependencies": {
     "@aws-sdk/credential-providers": "^3.201.0",
     "@mongodb-js/zstd": "^1.1.0",
+    "gcp-metadata": "^5.2.0",
     "kerberos": "^2.0.1",
     "mongodb-client-encryption": ">=2.3.0 <3",
-    "snappy": "^7.2.2",
-    "gcp-metadata": "^5.2.0"
+    "snappy": "^7.2.2"
   },
   "peerDependenciesMeta": {
     "@aws-sdk/credential-providers": {
@@ -93,6 +93,7 @@
     "js-yaml": "^4.1.0",
     "mocha": "^10.2.0",
     "mocha-sinon": "^2.1.2",
+    "mongodb-client-encryption": "^2.8.0",
     "mongodb-legacy": "^5.0.0",
     "nyc": "^15.1.0",
     "prettier": "^2.8.8",
@@ -135,7 +136,7 @@
     "check:dts": "node ./node_modules/typescript/bin/tsc --noEmit mongodb.d.ts && tsd",
     "check:search-indexes": "nyc mocha --config test/mocha_mongodb.json test/manual/search-index-management.spec.test.ts",
     "check:test": "mocha --config test/mocha_mongodb.json test/integration",
-    "check:unit": "mocha test/unit",
+    "check:unit": "mocha test/unit/client-side-encryption",
     "check:ts": "node ./node_modules/typescript/bin/tsc -v && node ./node_modules/typescript/bin/tsc --noEmit",
     "check:atlas": "mocha --config test/manual/mocharc.json test/manual/atlas_connectivity.test.js",
     "check:adl": "mocha --config test/mocha_mongodb.json test/manual/atlas-data-lake-testing",
