+++ test/integration/node-specific/client_encryption.test.ts
@@ -1,0 +1,552 @@
+import { expect } from 'chai';
+import { readFileSync } from 'fs';
+import * as sinon from 'sinon';
+
+import {
+  ClientEncryption,
+  type DataKey
+} from '../../../src/client-side-encryption/clientEncryption';
+import { StateMachine } from '../../../src/client-side-encryption/stateMachine';
+import { Binary, type Collection, Int32, Long, type MongoClient } from '../../mongodb';
+
+function readHttpResponse(path) {
+  let data = readFileSync(path, 'utf8').toString();
+  data = data.split('\n').join('\r\n');
+  return Buffer.from(data, 'utf8');
+}
+
+describe('ClientEncryption integration tests', function () {
+  let client: MongoClient;
+
+  beforeEach(async function () {
+    client = this.configuration.newClient();
+    await client.connect();
+    await client
+      .db('client')
+      .collection('encryption')
+      .drop()
+      .catch(() => null);
+  });
+
+  afterEach(async function () {
+    await client?.close();
+  });
+
+  describe('stubbed stateMachine', function () {
+    const sandbox = sinon.createSandbox();
+
+    after(() => sandbox.restore());
+    before(() => {
+      // stubbed out for AWS unit testing below
+      const MOCK_KMS_ENCRYPT_REPLY = readHttpResponse(
+        `${__dirname}/../../unit/client-side-encryption/data/kms-encrypt-reply.txt`
+      );
+      sandbox.stub(StateMachine.prototype, 'kmsRequest').callsFake(request => {
+        request.addResponse(MOCK_KMS_ENCRYPT_REPLY);
+        return Promise.resolve();
+      });
+    });
+
+    [
+      {
+        name: 'local',
+        kmsProviders: { local: { key: Buffer.alloc(96) } }
+      },
+      {
+        name: 'aws',
+        kmsProviders: { aws: { accessKeyId: 'example', secretAccessKey: 'example' } },
+        options: { masterKey: { region: 'region', key: 'cmk' } }
+      }
+    ].forEach(providerTest => {
+      it(`should create a data key with the "${providerTest.name}" KMS provider`, async function () {
+        const providerName = providerTest.name;
+        const encryption = new ClientEncryption(client, {
+          keyVaultNamespace: 'client.encryption',
+          kmsProviders: providerTest.kmsProviders
+        });
+
+        const dataKeyOptions = providerTest.options || {};
+
+        const dataKey = await encryption.createDataKey(providerName, dataKeyOptions);
+        expect(dataKey).property('_bsontype', 'Binary');
+
+        const doc = await client.db('client').collection('encryption').findOne({ _id: dataKey });
+        expect(doc).to.have.property('masterKey');
+        expect(doc.masterKey).property('provider', providerName);
+      });
+
+      it(`should create a data key with the "${providerTest.name}" KMS provider (fixed key material)`, async function () {
+        const providerName = providerTest.name;
+        const encryption = new ClientEncryption(client, {
+          keyVaultNamespace: 'client.encryption',
+          kmsProviders: providerTest.kmsProviders
+        });
+
+        const dataKeyOptions = {
+          ...providerTest.options,
+          keyMaterial: new Binary(Buffer.alloc(96))
+        };
+
+        const dataKey = await encryption.createDataKey(providerName, dataKeyOptions);
+        expect(dataKey).property('_bsontype', 'Binary');
+
+        const doc = await client.db('client').collection('encryption').findOne({ _id: dataKey });
+        expect(doc).to.have.property('masterKey');
+        expect(doc.masterKey).property('provider', providerName);
+      });
+    });
+
+    it(`should create a data key with the local KMS provider (fixed key material, fixed key UUID)`, async function () {
+      // 'Custom Key Material Test' prose spec test:
+      const keyVaultColl = client.db('client').collection('encryption');
+      const encryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders: {
+          local: {
+            key: 'A'.repeat(128) // the value here is not actually relevant
+          }
+        }
+      });
+
+      const dataKeyOptions = {
+        keyMaterial: new Binary(
+          Buffer.from(
+            'xPTAjBRG5JiPm+d3fj6XLi2q5DMXUS/f1f+SMAlhhwkhDRL0kr8r9GDLIGTAGlvC+HVjSIgdL+RKwZCvpXSyxTICWSXTUYsWYPyu3IoHbuBZdmw2faM3WhcRIgbMReU5',
+            'base64'
+          )
+        )
+      };
+      const dataKey = await encryption.createDataKey('local', dataKeyOptions);
+      expect(dataKey._bsontype).to.equal('Binary');
+
+      // Remove and re-insert with a fixed UUID to guarantee consistent output
+      const doc = (
+        await keyVaultColl.findOneAndDelete({ _id: dataKey }, { writeConcern: { w: 'majority' } })
+      ).value;
+      doc._id = new Binary(Buffer.alloc(16), 4);
+      await keyVaultColl.insertOne(doc, { writeConcern: { w: 'majority' } });
+
+      const encrypted = await encryption.encrypt('test', {
+        keyId: doc._id,
+        algorithm: 'AEAD_AES_256_CBC_HMAC_SHA_512-Deterministic'
+      });
+      expect(encrypted._bsontype).to.equal('Binary');
+      expect(encrypted.toString('base64')).to.equal(
+        'AQAAAAAAAAAAAAAAAAAAAAACz0ZOLuuhEYi807ZXTdhbqhLaS2/t9wLifJnnNYwiw79d75QYIZ6M/aYC1h9nCzCjZ7pGUpAuNnkUhnIXM3PjrA=='
+      );
+    });
+
+    it('should fail to create a data key if keyMaterial is wrong', function (done) {
+      const encryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders: { local: { key: 'A'.repeat(128) } }
+      });
+
+      const dataKeyOptions = {
+        keyMaterial: new Binary(Buffer.alloc(97))
+      };
+      try {
+        encryption.createDataKey('local', dataKeyOptions);
+        expect.fail('missed exception');
+      } catch (err) {
+        expect(err.message).to.equal('keyMaterial should have length 96, but has length 97');
+        done();
+      }
+    });
+
+    it('should explicitly encrypt and decrypt with the "local" KMS provider', async function () {
+      const encryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders: { local: { key: Buffer.alloc(96) } }
+      });
+
+      const dataKey = await encryption.createDataKey('local');
+
+      const encryptOptions = {
+        keyId: dataKey,
+        algorithm: 'AEAD_AES_256_CBC_HMAC_SHA_512-Deterministic'
+      };
+
+      const encrypted = await encryption.encrypt('hello', encryptOptions);
+      expect(encrypted._bsontype).to.equal('Binary');
+      expect(encrypted.sub_type).to.equal(6);
+
+      const decrypted = await encryption.decrypt(encrypted);
+      expect(decrypted).to.equal('hello');
+    });
+
+    it('should explicitly encrypt and decrypt with the "local" KMS provider (promise)', async function () {
+      const encryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders: { local: { key: Buffer.alloc(96) } }
+      });
+
+      const dataKey = await encryption.createDataKey('local');
+      const encryptOptions = {
+        keyId: dataKey,
+        algorithm: 'AEAD_AES_256_CBC_HMAC_SHA_512-Deterministic'
+      };
+
+      const encrypted = await encryption.encrypt('hello', encryptOptions);
+      expect(encrypted._bsontype).to.equal('Binary');
+      expect(encrypted.sub_type).to.equal(6);
+
+      const decrypted = await encryption.decrypt(encrypted);
+
+      expect(decrypted).to.equal('hello');
+    });
+
+    it('should explicitly encrypt and decrypt with a re-wrapped local key', async function () {
+      // Create new ClientEncryption instances to make sure
+      // that we are actually using the rewrapped keys and not
+      // something that has been cached.
+      const newClientEncryption = () =>
+        new ClientEncryption(client, {
+          keyVaultNamespace: 'client.encryption',
+          kmsProviders: { local: { key: 'A'.repeat(128) } }
+        });
+
+      const dataKey = await newClientEncryption().createDataKey('local');
+      const encryptOptions = {
+        keyId: dataKey,
+        algorithm: 'Indexed',
+        contentionFactor: 0
+      };
+
+      const encrypted = await newClientEncryption().encrypt('hello', encryptOptions);
+      expect(encrypted._bsontype).to.equal('Binary');
+      expect(encrypted.sub_type).to.equal(6);
+      const rewrapManyDataKeyResult = await newClientEncryption().rewrapManyDataKey({});
+      expect(rewrapManyDataKeyResult.bulkWriteResult.result.nModified).to.equal(1);
+      const decrypted = await newClientEncryption().decrypt(encrypted);
+    });
+
+    it('should not perform updates if no keys match', async function () {
+      const clientEncryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders: { local: { key: 'A'.repeat(128) } }
+      });
+
+      const rewrapManyDataKeyResult = await clientEncryption.rewrapManyDataKey({ _id: 12345 });
+      expect(rewrapManyDataKeyResult.bulkWriteResult).to.equal(undefined);
+    });
+
+    it.skip('should explicitly encrypt and decrypt with a re-wrapped local key (explicit session/transaction)', function () {
+      const encryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders: { local: { key: 'A'.repeat(128) } }
+      });
+      let encrypted;
+      let rewrapManyDataKeyResult;
+
+      return encryption
+        .createDataKey('local')
+        .then(dataKey => {
+          const encryptOptions = {
+            keyId: dataKey,
+            algorithm: 'Indexed',
+            contentionFactor: 0
+          };
+
+          return encryption.encrypt('hello', encryptOptions);
+        })
+        .then(_encrypted => {
+          encrypted = _encrypted;
+        })
+        .then(() => {
+          // withSession does not forward the callback's return value, hence
+          // the slightly awkward 'rewrapManyDataKeyResult' passing here
+          return client.withSession(session => {
+            return session.withTransaction(() => {
+              expect(session.transaction.isStarting).to.equal(true);
+              expect(session.transaction.isActive).to.equal(true);
+              rewrapManyDataKeyResult = encryption.rewrapManyDataKey(
+                {},
+                { provider: 'local', session }
+              );
+              return rewrapManyDataKeyResult.then(() => {
+                // Verify that the 'session' argument was actually used
+                expect(session.transaction.isStarting).to.equal(false);
+                expect(session.transaction.isActive).to.equal(true);
+              });
+            });
+          });
+        })
+        .then(() => {
+          return rewrapManyDataKeyResult;
+        })
+        .then(rewrapManyDataKeyResult => {
+          expect(rewrapManyDataKeyResult.bulkWriteResult.result.nModified).to.equal(1);
+          return encryption.decrypt(encrypted);
+        })
+        .then(decrypted => {
+          expect(decrypted).to.equal('hello');
+        });
+    }).skipReason = 'TODO(DRIVERS-2389): add explicit session support to key management API';
+
+    // TODO(NODE-3371): resolve KMS JSON response does not include string 'Plaintext'. HTTP status=200 error
+    it.skip('should explicitly encrypt and decrypt with the "aws" KMS provider', function (done) {
+      const encryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders: { aws: { accessKeyId: 'example', secretAccessKey: 'example' } }
+      });
+
+      const dataKeyOptions = {
+        masterKey: { region: 'region', key: 'cmk' }
+      };
+
+      encryption.createDataKey('aws', dataKeyOptions, (err, dataKey) => {
+        expect(err).to.not.exist;
+
+        const encryptOptions = {
+          keyId: dataKey,
+          algorithm: 'AEAD_AES_256_CBC_HMAC_SHA_512-Deterministic'
+        };
+
+        encryption.encrypt('hello', encryptOptions, (err, encrypted) => {
+          expect(err).to.not.exist;
+          expect(encrypted).to.have.property('v');
+          expect(encrypted.v._bsontype).to.equal('Binary');
+          expect(encrypted.v.sub_type).to.equal(6);
+
+          encryption.decrypt(encrypted, (err, decrypted) => {
+            expect(err).to.not.exist;
+            expect(decrypted).to.equal('hello');
+            done();
+          });
+        });
+      });
+    }).skipReason =
+      "TODO(NODE-3371): resolve KMS JSON response does not include string 'Plaintext'. HTTP status=200 error";
+  });
+
+  describe('encrypt()', function () {
+    let clientEncryption;
+    let completeOptions;
+    let dataKey;
+
+    beforeEach(async function () {
+      clientEncryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders: { local: { key: Buffer.alloc(96) } }
+      });
+
+      dataKey = await clientEncryption.createDataKey('local', {
+        name: 'local',
+        kmsProviders: { local: { key: Buffer.alloc(96) } }
+      });
+
+      completeOptions = {
+        algorithm: 'RangePreview',
+        contentionFactor: 0,
+        rangeOptions: {
+          min: new Long(0),
+          max: new Long(10),
+          sparsity: new Long(1)
+        },
+        keyId: dataKey
+      };
+    });
+
+    context('when expressionMode is incorrectly provided as an argument', function () {
+      it('overrides the provided option with the correct value for expression mode', async function () {
+        const optionsWithExpressionMode = { ...completeOptions, expressionMode: true };
+        const result = await clientEncryption.encrypt(new Long(0), optionsWithExpressionMode);
+
+        expect(result).to.be.instanceof(Binary);
+      });
+    });
+  });
+
+  describe('encryptExpression()', function () {
+    let clientEncryption;
+    let completeOptions;
+    let dataKey;
+    const expression = {
+      $and: [{ someField: { $gt: 1 } }]
+    };
+
+    beforeEach(async function () {
+      clientEncryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders: { local: { key: Buffer.alloc(96) } }
+      });
+
+      dataKey = await clientEncryption.createDataKey('local', {
+        name: 'local',
+        kmsProviders: { local: { key: Buffer.alloc(96) } }
+      });
+
+      completeOptions = {
+        algorithm: 'RangePreview',
+        queryType: 'rangePreview',
+        contentionFactor: 0,
+        rangeOptions: {
+          min: new Int32(0),
+          max: new Int32(10),
+          sparsity: new Long(1)
+        },
+        keyId: dataKey
+      };
+    });
+
+    it('throws if rangeOptions is not provided', async function () {
+      expect(delete completeOptions.rangeOptions).to.be.true;
+      const errorOrResult = await clientEncryption
+        .encryptExpression(expression, completeOptions)
+        .catch(e => e);
+
+      expect(errorOrResult).to.be.instanceof(TypeError);
+    });
+
+    it('throws if algorithm is not provided', async function () {
+      expect(delete completeOptions.algorithm).to.be.true;
+      const errorOrResult = await clientEncryption
+        .encryptExpression(expression, completeOptions)
+        .catch(e => e);
+
+      expect(errorOrResult).to.be.instanceof(TypeError);
+    });
+
+    it(`throws if algorithm does not equal 'rangePreview'`, async function () {
+      completeOptions['algorithm'] = 'equality';
+      const errorOrResult = await clientEncryption
+        .encryptExpression(expression, completeOptions)
+        .catch(e => e);
+
+      expect(errorOrResult).to.be.instanceof(TypeError);
+    });
+
+    it(`does not throw if algorithm has different casing than 'rangePreview'`, async function () {
+      completeOptions['algorithm'] = 'rAnGePrEvIeW';
+      const errorOrResult = await clientEncryption
+        .encryptExpression(expression, completeOptions)
+        .catch(e => e);
+
+      expect(errorOrResult).not.to.be.instanceof(Error);
+    });
+
+    context('when expressionMode is incorrectly provided as an argument', function () {
+      it('overrides the provided option with the correct value for expression mode', async function () {
+        const optionsWithExpressionMode = { ...completeOptions, expressionMode: false };
+        const result = await clientEncryption.encryptExpression(
+          expression,
+          optionsWithExpressionMode
+        );
+
+        expect(result).not.to.be.instanceof(Binary);
+      });
+    });
+  });
+
+  describe('ClientEncryptionKeyAltNames', function () {
+    let client: MongoClient;
+    let clientEncryption: ClientEncryption;
+    let collection: Collection<DataKey>;
+    // Data Key Stuff
+    const AWS_ACCESS_KEY_ID = process.env.AWS_ACCESS_KEY_ID;
+    const AWS_SECRET_ACCESS_KEY = process.env.AWS_SECRET_ACCESS_KEY;
+    const AWS_REGION = process.env.AWS_REGION;
+    const AWS_CMK_ID = process.env.AWS_CMK_ID;
+
+    // TODO - why are we using aws providers specifically?
+    const kmsProviders = {
+      aws: { accessKeyId: AWS_ACCESS_KEY_ID, secretAccessKey: AWS_SECRET_ACCESS_KEY }
+    };
+    const dataKeyOptions = { masterKey: { key: AWS_CMK_ID, region: AWS_REGION } };
+
+    // const SKIP_AWS_TESTS = [AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, AWS_REGION, AWS_CMK_ID].some(
+    //   secret => !secret
+    // );
+    beforeEach(function () {
+      client = this.configuration.newClient();
+      collection = client.db('client').collection('encryption');
+      clientEncryption = new ClientEncryption(client, {
+        keyVaultNamespace: 'client.encryption',
+        kmsProviders
+      });
+    });
+
+    afterEach(async function () {
+      await client?.close();
+    });
+
+    function makeOptions(keyAltNames) {
+      expect(dataKeyOptions.masterKey).to.be.an('object');
+      expect(dataKeyOptions.masterKey.key).to.be.a('string');
+      expect(dataKeyOptions.masterKey.region).to.be.a('string');
+
+      return {
+        masterKey: {
+          key: dataKeyOptions.masterKey.key,
+          region: dataKeyOptions.masterKey.region
+        },
+        keyAltNames
+      };
+    }
+
+    describe('errors', function () {
+      [42, 'hello', { keyAltNames: 'foobar' }, /foobar/].forEach(val => {
+        it(`should fail if typeof keyAltNames = ${typeof val}`, function () {
+          const options = makeOptions(val);
+          expect(() => clientEncryption.createDataKey('aws', options, () => undefined)).to.throw(
+            TypeError
+          );
+        });
+      });
+
+      [undefined, null, 42, { keyAltNames: 'foobar' }, ['foobar'], /foobar/].forEach(val => {
+        it(`should fail if typeof keyAltNames[x] = ${typeof val}`, function () {
+          const options = makeOptions([val]);
+          expect(() => clientEncryption.createDataKey('aws', options, () => undefined)).to.throw(
+            TypeError
+          );
+        });
+      });
+    });
+
+    it('should create a key with keyAltNames', async function () {
+      const options = makeOptions(['foobar']);
+      const dataKey = await clientEncryption.createDataKey('aws', options);
+      const document = await collection.findOne({ keyAltNames: 'foobar' });
+      expect(document).to.be.an('object');
+      expect(document).to.have.property('keyAltNames').that.includes.members(['foobar']);
+      expect(document).to.have.property('_id').that.deep.equals(dataKey);
+    });
+
+    it('should create a key with multiple keyAltNames', async function () {
+      const dataKey = await clientEncryption.createDataKey(
+        'aws',
+        makeOptions(['foobar', 'fizzbuzz'])
+      );
+      const docs = await Promise.all([
+        collection.findOne({ keyAltNames: 'foobar' }),
+        collection.findOne({ keyAltNames: 'fizzbuzz' })
+      ]);
+      expect(docs).to.have.lengthOf(2);
+      const doc1 = docs[0];
+      const doc2 = docs[1];
+      expect(doc1).to.be.an('object');
+      expect(doc2).to.be.an('object');
+      expect(doc1).to.have.property('keyAltNames').that.includes.members(['foobar', 'fizzbuzz']);
+      expect(doc1).to.have.property('_id').that.deep.equals(dataKey);
+      expect(doc2).to.have.property('keyAltNames').that.includes.members(['foobar', 'fizzbuzz']);
+      expect(doc2).to.have.property('_id').that.deep.equals(dataKey);
+    });
+
+    it('should be able to reference a key with `keyAltName` during encryption', async function () {
+      const keyAltName = 'mySpecialKey';
+      const algorithm = 'AEAD_AES_256_CBC_HMAC_SHA_512-Deterministic';
+
+      const valueToEncrypt = 'foobar';
+
+      const keyId = await clientEncryption.createDataKey('aws', makeOptions([keyAltName]));
+      const encryptedValue = await clientEncryption.encrypt(valueToEncrypt, { keyId, algorithm });
+      const encryptedValue2 = await clientEncryption.encrypt(valueToEncrypt, {
+        keyAltName,
+        algorithm
+      });
+      expect(encryptedValue).to.deep.equal(encryptedValue2);
+    });
+  });
+});
